Server/requirements.txt
asgiref==3.3.1
coverage==5.5
Django==3.1.7
django-cors-headers==3.7.0
django-nose==1.4.7
django-rest==0.8.6
djangorestframework==3.12.2
docutils==0.17
nose==1.3.7
pytz==2021.1
six==1.12.0
sqlparse==0.4.1
